requirements.txt
# LiveKit core
livekit-agents[openai,deepgram,cartesia,silero,turn_detector,google,groq,elevenlabs]~=1.3
livekit-plugins-assemblyai

livekit-plugins-noise-cancellation~=0.2

# Memory / tools
mem0ai
duckduckgo-search
ddgs
langchain_community
requests

# Backend
fastapi
uvicorn[standard]
python-multipart

# DB
pymongo>=4.6.0
motor>=3.3.0

# Utils
python-dotenv==1.1.1
pytz==2025.2
edge-tts

# Google Calendar
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
